# docs/openapi.yaml
openapi: 3.1.0
info:
  title: KISY Enterprise Messenger API
  version: 0.4.0
  description: >
    REST API for the KISY enterprise messenger. Covers stages 2–4:
    authentication, access control, invitations, profile, groups, private
    chats, messages, reactions, read state, favorites, notifications, the
    CEO admin surface and the WebSocket gateway. File handling is added in a
    later stage.
servers:
  - url: /api/v1
    description: Versioned API root (behind the Nginx edge proxy)
...
paths:
...
  /chats:
    get:
      tags: [chats]
      summary: List the current user's private chats
      responses:
        '200':
          description: Private chats
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data:
                        type: object
                        properties:
                          chats:
                            type: array
                            items: { $ref: '#/components/schemas/Chat' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [chats]
      summary: Open (or fetch) a private chat with another user
      description: >
        Idempotent. Enforces the initiation rule: a user may open a chat
        only with same-or-lower clearance; the existing chat is returned if
        one already exists.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId: { type: string, format: uuid }
      responses:
        '200':
          description: Chat opened or already existed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data:
                        type: object
                        properties:
                          chat: { $ref: '#/components/schemas/Chat' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /chats/{chatID}:
    get:
      tags: [chats]
      summary: Get a private chat the user participates in
      parameters:
        - { name: chatID, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Chat
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data:
                        type: object
                        properties:
                          chat: { $ref: '#/components/schemas/Chat' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /favorites:
    get:
      tags: [chats]
      summary: List favorite/pinned chats
      responses:
        '200':
          description: Favorites, pinned first
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data:
                        type: object
                        properties:
                          favorites:
                            type: array
                            items: { $ref: '#/components/schemas/Favorite' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    put:
      tags: [chats]
      summary: Add or update a favorite/pin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [chatType, chatId]
              properties:
                chatType: { type: string, enum: [private, group] }
                chatId: { type: string, format: uuid }
                isPinned: { type: boolean }
                pinnedOrder: { type: [integer, 'null'] }
      responses:
        '200': { description: Stored, content: { application/json: { schema: { $ref: '#/components/schemas/Envelope' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [chats]
      summary: Remove a favorite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [chatType, chatId]
              properties:
                chatType: { type: string, enum: [private, group] }
                chatId: { type: string, format: uuid }
      responses:
        '200': { description: Removed, content: { application/json: { schema: { $ref: '#/components/schemas/Envelope' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /admin/users:
    get:
      tags: [admin]
      summary: List accounts (CEO only)
      parameters:
        - { name: limit, in: query, required: false, schema: { type: integer, default: 50, maximum: 200 } }
        - { name: offset, in: query, required: false, schema: { type: integer, default: 0 } }
      responses:
        '200':
          description: Users
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data:
                        type: object
                        properties:
                          users:
                            type: array
                            items: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/users/{userID}/role:
    patch:
      tags: [admin]
      summary: Change a user's clearance level (CEO only)
      parameters:
        - { name: userID, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [roleLevel]
              properties:
                roleLevel: { type: integer, minimum: 1, maximum: 10 }
      responses:
        '200': { description: Role changed, content: { application/json: { schema: { $ref: '#/components/schemas/Envelope' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/users/{userID}/reset-password:
    post:
      tags: [admin]
      summary: Reset a user's password (CEO only); revokes their sessions
      parameters:
        - { name: userID, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [newPassword]
              properties:
                newPassword: { type: string, minLength: 12, maxLength: 128 }
      responses:
        '200': { description: Password reset, content: { application/json: { schema: { $ref: '#/components/schemas/Envelope' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/users/{userID}/activate:
    post:
      tags: [admin]
      summary: Activate a user (CEO only)
      parameters:
        - { name: userID, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Activated, content: { application/json: { schema: { $ref: '#/components/schemas/Envelope' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/users/{userID}/deactivate:
    post:
      tags: [admin]
      summary: Deactivate a user (CEO only); revokes their sessions
      parameters:
        - { name: userID, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Deactivated, content: { application/json: { schema: { $ref: '#/components/schemas/Envelope' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/audit:
    get:
      tags: [admin]
      summary: Read the audit log (CEO only)
      parameters:
        - { name: action, in: query, required: false, schema: { type: string } }
        - { name: limit, in: query, required: false, schema: { type: integer, default: 50 } }
        - { name: offset, in: query, required: false, schema: { type: integer, default: 0 } }
      responses:
        '200':
          description: Audit entries, newest first
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data:
                        type: object
                        properties:
                          entries:
                            type: array
                            items: { $ref: '#/components/schemas/AuditEntry' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
components:
  securitySchemes:
    accessCookie:
      type: apiKey
      in: cookie
      name: kisy_access
    refreshCookie:
      type: apiKey
      in: cookie
      name: kisy_refresh
...
  schemas:
    Envelope:
      type: object
      required: [success, requestId, timestamp]
      properties:
        success: { type: boolean }
        data: { type: object }
        error: { $ref: '#/components/schemas/Error' }
        requestId: { type: string }
        timestamp: { type: string, format: date-time }
    Error:
      type: object
      properties:
        code:
          type: string
          enum:
            - AUTH_INVALID_TOKEN
            - AUTH_INVALID_CREDENTIALS
            - AUTH_EXPIRED
            - ACCESS_DENIED
            - RESOURCE_NOT_FOUND
            - RATE_LIMITED
            - VALIDATION_FAILED
            - INTERNAL_ERROR
        message: { type: string }
        details: {}
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        displayName: { type: string }
        roleLevel: { type: integer, minimum: 1, maximum: 10 }
        avatarUrl: { type: [string, 'null'] }
        status: { type: string, enum: [online, offline, away] }
        lastSeen: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
...
    Chat:
      type: object
      properties:
        id: { type: string, format: uuid }
        type: { type: string, enum: [private] }
        otherUserId: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
...
    Favorite:
      type: object
      properties:
        chatType: { type: string, enum: [private, group] }
        chatId: { type: string, format: uuid }
        isPinned: { type: boolean }
        pinnedOrder: { type: [integer, 'null'] }
...
    AuditEntry:
      type: object
      properties:
        id: { type: string, format: uuid }
        actorId: { type: [string, 'null'], format: uuid }
        action: { type: string }
        targetType: { type: [string, 'null'] }
        targetId: { type: [string, 'null'], format: uuid }
        requestId: { type: [string, 'null'] }
        metadata: { type: object }
        createdAt: { type: string, format: date-time }
  responses:
    ValidationError:
      description: Request failed validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
    Unauthorized:
      description: Authentication required or failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
    Forbidden:
      description: Insufficient clearance
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
    NotFound:
      description: Resource not found (or hidden from this clearance)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
...
security:
  - accessCookie: []